JBCOOL-9 #progress #time 3h : Clean up the code for REST, now xml and json services are delivered by the same object CoolRESTImpl and implement a common interface. Some cleanup in logging messages as well. Reset the OrderBy once it has been applied.

+++ JBRestCool/src/atlas/cool/jsf/beans/CoolSchemaParamsBean.java
@@ -322,13 +322,13 @@ public class CoolSchemaParamsBean implements Serializable {
 					+ " " + selNode.getNodeFullpath() + " " + channelId + " "
 					+ stime + " " + etime);
 
-			Integer chId = null;
+			Long chId = null;
 
 //			if (channelId.equals("") || channelId.equals("all")) {
 			if (channelId != null && channelId.equals("all")) {
 				log.info("no cut on channel ID...");
 			} else {
-				chId = selChannel.getChannelId().intValue();
+				chId = selChannel.getChannelId().longValue();
 			}
 //				chId = new Integer(channelId);
